Linux compatibility,bugs fixed
-freq and voltages settings dialog:
 -shows multiplier, frequeny in MHz and voltage in Volt from the slider values in the "voltage and frequency settings" dialog after showing/ opening the dialog
-mainframe:
 -OnSaveAsDefaultPStates(...): may create p-state dir path if it does not exist
 -draws available voltages and frequencies as much as the space allows
-bugs fixed:
 -after detaching/ unloading CPU cotroller (dynamic library) the wrong menu ("Extras") is deleted instead of the "core(s)" menu
 -the main frame does not draw anything when loading/ attaching a CPU controller dynamic library manually
 -p-state cross in main frame goes down (instead of upwards) when the voltage is increased via the "voltage and frequency settings" dialog
 -made the checkbox "also set wanted voltage" in the "voltage and frequency settings" dialog working

+++ wxWidgets/UserInterface/CPUregisterWriteDialog.hpp
@@ -24,6 +24,7 @@
 #include <wx/dialog.h>
 #include <vector>
 
+//Forward declaration instead of include's(=faster)
 class Model ;
 class I_CPUcontroller ;
 class MSRdata ;
@@ -43,7 +44,7 @@ private:
   wxGridSizer * mp_wxgridsizerAttributeNameAndValue ;
   wxListBox * p_wxlistbox ;
   std::vector<wxTextCtrl * > m_stdvec_p_wxtextctrl ;
-  WORD m_wNumIgnoreChanges ;
+  unsigned long m_wNumIgnoreChanges ;
 public:
   CPUregisterWriteDialog(
     wxWindow * parent ,
